updated event entity and respository

+++ ultragamecenter/src/main/java/urjc/ugc/ultragamecenter/EventRepository.java
@@ -4,5 +4,5 @@ package urjc.ugc.ultragamecenter;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EventRepository extends JpaRepository<Event, Integer> {
-    Event findByid(int id);
+    Event findByid(long id);
 }
